requirements.txt
apache-airflow-providers-postgres
psycopg2-binary  # DB-API driver for PostgreSQL
huggingface-hub  # Data source